+++ xpeapp_android/app/src/main/java/com/xpeho/xpeapp/ui/componants/newsletter/NewsletterCard.kt
@@ -1,6 +1,7 @@
 package com.xpeho.xpeapp.ui.componants.newsletter
 
 import androidx.compose.foundation.background
+import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
@@ -10,6 +11,7 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.outlined.ArrowForwardIos
 import androidx.compose.material.icons.filled.Newspaper
 import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -34,7 +36,8 @@ fun NewsletterCard(newsletter: Newsletter) {
             .height(60.dp)
             .padding(10.dp)
             .background(
-                color = Color.White,
+                color = if (isSystemInDarkTheme())
+                    MaterialTheme.colorScheme.surfaceVariant else Color.White,
                 shape = RoundedCornerShape(50.dp)
             )
     ) {
@@ -58,6 +61,7 @@ fun NewsletterCard(newsletter: Newsletter) {
             ),
             fontFamily = SfPro,
             fontWeight = FontWeight.Bold,
+            color = MaterialTheme.colorScheme.onSurface
         )
         Spacer(
             modifier = Modifier.weight(1f)